Move logic related to Trip Type select and Passengers out to the SearchFormModes to separate components
